frontend/requirements.txt

requests==2.28.2
numpy==1.24.2 
Pillow==9.4.0 
matplotlib==3.7.0 
